requirements.txt
qdrant_client
sentence_transformers
openai
python-dotenv